Finished parsing Map Objects but get two Unrecognized Object ID 199, 50

+++ Sources/Backend/HoMM3/Models/Faction.swift
@@ -6,7 +6,8 @@
 //
 
 import Foundation
-public enum Faction: Int, Equatable, CaseIterable, CustomStringConvertible {
+
+public enum Faction: UInt8, Equatable, CaseIterable, CustomStringConvertible {
     case castle, rampart, tower, inferno, necropolis, dungeon, stronghold, fortress, conflux, neutral
 }
